Move diagnostic into own file
