+++ crates/compiler/src/listeners/compiler_listener.rs
@@ -1,6 +1,9 @@
 use crate::prelude::*;
+use antlr_rust::tree::ParseTreeListener;
 use rusty_yarn_spinner_core::prelude::*;
 mod emit;
+use crate::parser::generated::yarnspinnerparser::YarnSpinnerParserContextType;
+use crate::prelude::generated::yarnspinnerparserlistener::YarnSpinnerParserListener;
 pub(crate) use emit::*;
 
 #[derive(Debug)]
@@ -10,3 +13,7 @@ pub(crate) struct CompilerListener {
     /// The current debug information that describes [`current_node`].
     pub(crate) current_debug_info: DebugInfo,
 }
+
+impl<'input> ParseTreeListener<'input, YarnSpinnerParserContextType> for CompilerListener {}
+
+impl<'input> YarnSpinnerParserListener<'input> for CompilerListener {}
